Add unit tests for untested new functionality
