+++ p2p/sentinel.go
@@ -3,7 +3,6 @@ package p2p
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 	"net/http"
 
 	"github.com/tendermint/tendermint/libs/log"
@@ -12,6 +11,26 @@ import (
 func RegisterWithSentinel(logger log.Logger, APIKey, validatorAddrHex, peerID, sentinel string) {
 	logger.Info("[p2p.sentinel]: Registering with sentinel", APIKey, validatorAddrHex, peerID, sentinel)
 
+	jsonData, err := makePostRequestData(APIKey, validatorAddrHex, peerID)
+	if err != nil {
+		logger.Info("[p2p.sentinel]: Err marshalling json data:", err)
+		return
+	}
+
+	go postRequestRoutine(logger, sentinel, jsonData)
+}
+
+func postRequestRoutine(logger log.Logger, sentinel string, jsonData []byte) {
+	resp, err := http.Post(sentinel, "application/json", bytes.NewBuffer(jsonData)) //nolint:gosec
+	if err != nil {
+		logger.Info("[p2p.sentinel]: Err making post request to sentinel:", err)
+	} else {
+		logger.Info("[p2p.sentinel]: Successfully registered with sentinel", resp)
+		defer resp.Body.Close()
+	}
+}
+
+func makePostRequestData(APIKey, validatorAddrHex, peerID string) ([]byte, error) {
 	params := [3]string{APIKey, validatorAddrHex, peerID}
 	data := map[string]interface{}{
 		"method": "register_peer",
@@ -19,19 +38,5 @@ func RegisterWithSentinel(logger log.Logger, APIKey, validatorAddrHex, peerID, s
 		"id":     1,
 	}
 
-	jsonData, err := json.Marshal(data)
-	if err != nil {
-		fmt.Println("[p2p.sentinel]: Err marshalling json data:", err)
-		return
-	}
-
-	go func() {
-		resp, err := http.Post(sentinel, "application/json", bytes.NewBuffer(jsonData)) //nolint:gosec
-		if err != nil {
-			logger.Info("[p2p.sentinel]: Err making post request to sentinel:", err)
-		} else {
-			logger.Info("[p2p.sentinel]: Successfully registered with sentinel", resp)
-			defer resp.Body.Close()
-		}
-	}()
+	return json.Marshal(data)
 }
